requirements.txt
fastapi
uvicorn
python-binance
kiteconnect
angelbroking-smartapi
python-dotenv
logzero
pyotp
websocket-client
jinja2